+++ app/presenters/refund_presenter.rb
@@ -1,0 +1,6 @@
+class RefundPresenter < SimpleDelegator
+  def value
+    return if super.blank?
+    ActionController::Base.helpers.number_to_currency(super, unit: "£")
+  end
+end